requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0
google-genai>=1.0.0
reportlab>=4.0.0
